상태창에 dday 표시 기능 구현

+++ app/src/main/java/kr/ricale/dday/utils/ImageUtil.kt
@@ -13,6 +13,10 @@ import java.io.FileOutputStream
 object ImageUtil {
     private lateinit var context: Context
 
+    fun init(c: Context) {
+        context = c
+    }
+
     private fun getFile(dirname: String, filename: String): File {
         val cw = ContextWrapper(context)
         val directory = cw.getDir(dirname, Context.MODE_PRIVATE)
@@ -60,12 +64,11 @@ object ImageUtil {
         )
     }
 
-    fun init(c: Context) {
-        context = c
-    }
-
     fun getImage(dirname: String, filename: String): Bitmap? {
         val file = getFile(dirname, filename)
+        if(!file.isFile) {
+            return null
+        }
 
         val options = BitmapFactory.Options()
         options.inPreferredConfig = Bitmap.Config.ARGB_8888
@@ -81,7 +84,9 @@ object ImageUtil {
 
     fun removeImage(dirname: String, filename: String) {
         val file = getFile(dirname, filename)
-        file.delete()
+        if(file.isFile) {
+            file.delete()
+        }
     }
 
     fun getImageFromUri(imageUri: Uri): Bitmap {
